backend/pyproject.toml
[tool.poetry]
name = "personal-crm-backend"
version = "0.1.0"
description = ""
authors = ["Chris Watkins <chris.watkins93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytz = "^2021.1"
psycopg2-binary = "^2.9.1"
SQLAlchemy = "1.3.24"
cffi = "^1.15.0"
python-dateutil = "^2.8.2"
fastapi = "^0.70.1"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
alembic = "^1.7.5"
click = "^8.0.3"

[tool.poetry.dev-dependencies]
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
